+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Tx.hs
@@ -61,9 +61,9 @@ import Cardano.Ledger.Binary (
 import Cardano.Ledger.Binary.Coders
 import qualified Cardano.Ledger.Binary.Plain as Plain
 import Cardano.Ledger.Coin (Coin)
-import Cardano.Ledger.Credential (Credential (..))
+import Cardano.Ledger.Credential (Credential (..), credKeyHashWitness)
 import Cardano.Ledger.Crypto (Crypto, StandardCrypto)
-import Cardano.Ledger.Keys (HasKeyRole (coerceKeyRole), KeyHash, KeyRole (Witness), asWitness)
+import Cardano.Ledger.Keys (HasKeyRole (coerceKeyRole), KeyHash, KeyRole (Witness))
 import Cardano.Ledger.Keys.Bootstrap (bootstrapWitKeyHash)
 import Cardano.Ledger.Keys.WitVKey (witVKeyHash)
 import Cardano.Ledger.MemoBytes (Mem, MemoBytes, memoBytes, mkMemoBytes, pattern Memo)
@@ -79,8 +79,8 @@ import Cardano.Ledger.Val ((<+>), (<×>))
 import Control.DeepSeq (NFData)
 import qualified Data.ByteString.Lazy as LBS
 import qualified Data.ByteString.Short as SBS
-import Data.Foldable as F (foldMap')
 import Data.Map.Strict (Map)
+import Data.Maybe (mapMaybe)
 import Data.Maybe.Strict (
   StrictMaybe (..),
   maybeToStrictMaybe,
@@ -399,8 +399,8 @@ extractKeyHashWitnessSet ::
   forall (r :: KeyRole) c.
   [Credential r c] ->
   Set (KeyHash 'Witness c)
-extractKeyHashWitnessSet =
-  F.foldMap' (maybe mempty Set.singleton . toKeyHashWitness)
+extractKeyHashWitnessSet = Set.fromList . mapMaybe credKeyHashWitness
+{-# DEPRECATED extractKeyHashWitnessSet "In favor of `credKeyHashWitness`" #-}
 
 -- | Minimum fee calculation
 shelleyMinFeeTx :: EraTx era => PParams era -> Tx era -> Coin